core/requirements.txt
fastapi==0.115.12
httpx==0.27.0
langchain==0.3.20
langchain-ollama==0.2.3
langchain-tavily==0.1.5
langgraph==0.3.5
python-dotenv==1.0.1
pymysql==1.1.1
django==5.1.7
uvicorn==0.34.0
wikipedia==1.4.0
